change data paths and column reordering

+++ tests/test_benchmark.py
@@ -51,7 +51,7 @@ def test__sort_leaderboard_no_rank():
     score = {k: range(5) for k in metrics.keys()}
 
     score['pipeline'] = range(5)
-    score = pd.DataFrame(score, dtype='int64')
+    score = pd.DataFrame(score)
 
     expected_return = score.iloc[::-1].reset_index(drop=True)
     expected_return['rank'] = range(1, 6)
